refactor: share fsReason and ReportError from src/report/errors.ts
Addresses review feedback on #19: fsReason was copied verbatim into
init.ts, html.ts and junit.ts. Move it (and ReportError, which html.ts
already imported from junit.ts) into a shared src/report/errors.ts so
there is one definition. No behavior change.
