+++ src/report/junit.ts
@@ -2,6 +2,7 @@ import { mkdir, writeFile } from 'node:fs/promises';
 import path from 'node:path';
 import type { BudgetSpend } from '../runner/budget.js';
 import type { TestResult } from '../runner/executor.js';
+import { ReportError, fsReason } from './errors.js';
 
 const ESCAPES: Record<string, string> = {
   '&': '&amp;',
@@ -117,26 +118,6 @@ export function renderJUnit(
 
   lines.push('</testsuite>', '');
   return lines.join('\n');
-}
-
-export class ReportError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = 'ReportError';
-  }
-}
-
-/**
- * Reduces a Node fs error to plain prose: drops the leading errno code
- * (`EACCES:`, `EISDIR:`, …) and the trailing `, <syscall> '<path>'` (the path
- * is already named in the surrounding message). Non-Error throws pass through.
- */
-function fsReason(error: unknown): string {
-  if (!(error instanceof Error)) return String(error);
-  return error.message
-    .replace(/^E[A-Z]+:\s*/, '')
-    .replace(/,\s+\w+\s+'[^']*'$/, '')
-    .trim();
 }
 
 /** Writes the report, creating missing parent directories. Returns the path written. */
